chore: moves operator/subscriptions operations to pkg/cluster
They do not belong to pkg/deploy as they are about reading/writing cluster resources rather than deploying resources.
